[Feat] Inventory entity 수정

+++ src/main/java/com/sanapyeong/mtvs_3rd_dreamplanet/Inventory/repositories/InventoryRepository.java
@@ -4,5 +4,5 @@ package com.sanapyeong.mtvs_3rd_dreamplanet.Inventory.repositories;
 import com.sanapyeong.mtvs_3rd_dreamplanet.Inventory.entities.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
+public interface InventoryRepository extends JpaRepository<Inventory, Long> {
 }
